requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
charset-normalizer==2.1.1
Django==4.1.5
ffmpeg-python==0.2.0
frozenlist==1.3.3
future==0.18.3
idna==3.4
imageio-ffmpeg==0.4.8
multidict==6.0.4
numpy==1.24.1
py-cord==2.3.2
python-dotenv==0.21.1
sqlparse==0.4.3
typing_extensions==4.4.0
tzdata==2022.7
yarl==1.8.2
